+++ workalendar/exceptions.py
@@ -19,3 +19,21 @@ class ISORegistryError(CalendarError):
     """
     Raised when you are trying to register a non-Calendar object
     """
+
+
+class ICalExportError(CalendarError):
+    """
+    Raised when the iCal export is impossible
+    """
+
+
+class ICalExportRangeError(ICalExportError):
+    """
+    Raised when the iCal export is impossible due to wrong range
+    """
+
+
+class ICalExportTargetPathError(ICalExportError):
+    """
+    Raised when the iCal export is impossible due to unusable target path
+    """