+++ Changelog.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## master (unreleased)
+
+Nothing here yet.
+
 ## v10.4.0 (2020-08-28)
 
 ### New calendar