+++ Changelog.md
@@ -4,6 +4,12 @@
 
 Nothing here yet.
 
+
+### New calendar
+
+- Added Mozambique calendar by @mr-shovel ().
+
+
 ## v10.4.0 (2020-08-28)
 
 ### New calendar