+++ Changelog.md
@@ -6,6 +6,9 @@
 
 - Added Mozambique calendar by @mr-shovel (#542).
 
+### New feature
+
+- Added iCal export feature, initiated by @joooeey (#197).
 
 ## v10.4.0 (2020-08-28)